Fixed problem with unicode encoding.
When raw results are filtered, they are returned as unicode
strings.  The system can now simply work with unicode strings and
not have to worry about encoding/decoding to be compatible with django
ORM.

Handles breakpoints was adjusted to work with unicode rather than byte
strings.

Runner model repr method returns a byte string which causes shell to choke
when trying to decode with ascii codec.  Changed repr to unicode method to
explicitly tell shell contents are unicode.