FindsOrCreatesResult utility.
FindsOrCreatesResult is an extension of FindsOrCreatesObjects. This
utility is specific for finding Result objects and only requires the
Runner and Event attributes to match for a successful find.

Whether it finds an existing Result in this fashion, or creates a new
one, the rest of the attributes are mapped to the returned Result object.

The motivation for this is to allow easy corrections to existing results.
Feeding a set of results through this will essentially perform an update
operation to any existing results, and a create to those which do not already
exist.